Fix a crash on about screen on iOS

+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/about/AboutScreen.kt
@@ -61,7 +61,7 @@ fun AboutScreen(
         text = stringResource(
             MR.strings.version,
             versionName,
-            versionCode.toInt()
+            versionCode
         ),
     )
   }
